requirements.txt
streamlit
streamlit_extras
langchain>=0.1.0
langchain-community>=0.0.20
langchain-google-genai>=0.0.11
langchain-core>=0.1.0
pydantic>=2.0.0
numpy
faiss-cpu
sentence-transformers
google-generativeai
PyPDF2
python-dotenv